Add files via upload
Program A:
This program asks the user to enter a message. It sends the message to Program B and prints the uppercase response it gets back.
Program B:
This program waits for a connection from Program A. When it receives a message, it changes the message to uppercase and sends it back.